CPLGetPID() now returns GIntBig (#1963)
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@13253 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/nitf/nitfbilevel.cpp
@@ -57,7 +57,7 @@ int NITFUncompressBILEVEL( NITFImage *psImage,
 /* -------------------------------------------------------------------- */
     CPLString osFilename;
 
-    osFilename.Printf( "/vsimem/nitf-wrk-%d.tif", CPLGetPID() );
+    osFilename.Printf( "/vsimem/nitf-wrk-%ld.tif", (long) CPLGetPID() );
 
     TIFF *hTIFF = VSI_TIFFOpen( osFilename, "w+" );
